Spatial regularization fix
Spatial regularization wasn't handling the edges of the image correctly.
I've implemented the system by shifting the
weight-vector-thought-of-as-a-28-x-28-matrix in each of the four compass
directions, by either shifting the vector up or down 1 or 28 places
(populating the vacated cells with 0s).  I avoided reshaping the vector
because I couldn't find any library functions that simplified the task
either way, and this section of code is speed critical.  The 1 shifts
(which represent left or right shifts of the matrix) create a
wrap-around effect that this update fixes.